zend_parse_parameters l must be long
git-svn-id: https://svn.php.net/repository/pecl/eio/trunk@331729 c90b9560-bf6c-de11-be94-00142212c4b1

+++ php_eio.c
@@ -2382,12 +2382,12 @@ PHP_FUNCTION(eio_set_max_poll_time)
 #define EIO_SET_INT_FUNCTION(eio_func) \
 	PHP_FUNCTION(eio_func) \
 	{ \
-		unsigned int num; \
+		unsigned long num; \
 		\
 		if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "l", &num) == FAILURE) { \
 			return; \
 		} \
-		eio_func(num); \
+		eio_func((unsigned int)num); \
 	}
 
 #define EIO_GET_INT_FUNCTION(eio_func) \
